+++ kernel/common/core/Cargo.toml
@@ -1,0 +1,3 @@
+[package]
+name = "ructiss_core"
+version = "0.0.1"
